feat: add DestinatarioZonePicker and TagZonePicker with multi-variant support
Create proper picker components matching the CategoryZonePicker pattern:
- Auto-switch between popover (desktop) and drawer (mobile) via useMediaQuery
- Lazy-load data on first open, cache in state for subsequent opens
- Search, selection, and inline tag creation

Wire into movimientos transaction row, replacing the custom drawer components.
Delete the now-unused destinatario-drawer.tsx and tag-drawer.tsx.
